+++ enemies/Normie.java
@@ -1,0 +1,12 @@
+package enemies;
+
+import java.awt.Color;
+import physics.*;
+
+public class Normie extends Enemy {
+
+    public Normie(Vector2 pos) {
+        super(pos, Color.red, 20);
+    }
+
+}